+++ Tests/PulseTests/PersistenLogHandlerTests.swift
@@ -8,23 +8,25 @@ import Logging
 @testable import Pulse
 
 final class PersistentLogHandlerTests: XCTestCase {
+    var tempDirectoryURL: URL!
+    var storeURL: URL!
+
+    var store: LoggerMessageStore!
+    var sut: PersistentLogHandler!
+
     override func setUp() {
-        super.setUp()
+        tempDirectoryURL = FileManager().temporaryDirectory.appendingPathComponent(UUID().uuidString)
+        try? FileManager.default.createDirectory(at: tempDirectoryURL, withIntermediateDirectories: true, attributes: [:])
+        storeURL = tempDirectoryURL.appendingPathComponent("test-store")
+
+        store = LoggerMessageStore(storeURL: storeURL)
+
+        sut = PersistentLogHandler(label: "test-hanlder", store: store, makeCurrentDate: { Date() })
     }
 
     override func tearDown() {
-        super.tearDown()
-
-        LoggerMessageStore.default.removeAllMessages()
-    }
-
-    func testItStoresMetadata() throws {
-        let id = UUID()
-
-        var logger = Logger(label: "test.logger", factory: { PersistentLogHandler(label: $0) })
-        logger[metadataKey: "id"] = "\(id)"
-
-        XCTAssertEqual(logger[metadataKey: "id"], "\(id)")
+        store.destroyStores()
+        try? FileManager.default.removeItem(at: tempDirectoryURL)
     }
 
     func testItPersistsLoggedMessages() throws {
@@ -41,7 +43,7 @@ final class PersistentLogHandlerTests: XCTestCase {
 
         LoggingSystem.bootstrap {
             MultiplexLogHandler([
-                PersistentLogHandler(label: $0, store: LoggerMessageStore.default, makeCurrentDate: { date }),
+                PersistentLogHandler(label: $0, store: self.store, makeCurrentDate: { date }),
                 StreamLogHandler.standardOutput(label: $0)
             ])
         }
@@ -59,21 +61,83 @@ final class PersistentLogHandlerTests: XCTestCase {
 
         waitForExpectations(timeout: 0.5)
 
-        let persistedMessages = try LoggerMessageStore.default.allMessages()
+        let persistedMessages = try store.allMessages()
         XCTAssertEqual(persistedMessages.count, 2)
 
-        let persistedMessage1 = persistedMessages[0]
-        XCTAssertEqual(persistedMessage1.level, level1.rawValue)
-        XCTAssertEqual(persistedMessage1.text, message1)
-        XCTAssertEqual(persistedMessage1.createdAt, date)
-        XCTAssertEqual(persistedMessage1.label, "test.logger.1")
-        XCTAssertEqual(persistedMessage1.session, sessionID.uuidString)
+        if persistedMessages.count >= 2 {
+            let persistedMessage1 = persistedMessages[0]
+            XCTAssertEqual(persistedMessage1.level, level1.rawValue)
+            XCTAssertEqual(persistedMessage1.text, message1)
+            XCTAssertEqual(persistedMessage1.createdAt, date)
+            XCTAssertEqual(persistedMessage1.label, "test.logger.1")
+            XCTAssertEqual(persistedMessage1.session, sessionID.uuidString)
 
-        let persistedMessage2 = persistedMessages[1]
-        XCTAssertEqual(persistedMessage2.level, level2.rawValue)
-        XCTAssertEqual(persistedMessage2.text, message2)
-        XCTAssertEqual(persistedMessage2.createdAt, date)
-        XCTAssertEqual(persistedMessage2.label, "test.logger.2")
-        XCTAssertEqual(persistedMessage2.session, sessionID.uuidString)
+            let persistedMessage2 = persistedMessages[1]
+            XCTAssertEqual(persistedMessage2.level, level2.rawValue)
+            XCTAssertEqual(persistedMessage2.text, message2)
+            XCTAssertEqual(persistedMessage2.createdAt, date)
+            XCTAssertEqual(persistedMessage2.label, "test.logger.2")
+            XCTAssertEqual(persistedMessage2.session, sessionID.uuidString)
+        }
+    }
+
+    // MARK: Metadata
+
+    func testStoringStringMetadata() throws {
+        // WHEN
+        sut.log(level: .debug, message: "request failed", metadata: ["system": "auth"])
+        flush(store: store)
+
+        // THEN key-value metadata is stored
+        let message = try XCTUnwrap(store.allMessages().first)
+        XCTAssertEqual(message.metadata.count, 1)
+        let entry = try XCTUnwrap(message.metadata.first)
+        XCTAssertEqual(entry.key, "system")
+        XCTAssertEqual(entry.value, "auth")
+    }
+
+    func testStoringStringConvertibleMetadata() throws {
+        // GIVEN
+        struct Foo: CustomStringConvertible {
+            var description: String {
+                return "foo"
+            }
+        }
+
+        // WHEN
+        sut.log(level: .debug, message: "a", metadata: ["system": .stringConvertible(Foo())])
+        flush(store: store)
+
+        // THEN key-value metadata is stored
+        let message = try XCTUnwrap(store.allMessages().first)
+        XCTAssertEqual(message.metadata.count, 1)
+        let entry = try XCTUnwrap(message.metadata.first)
+        XCTAssertEqual(entry.key, "system")
+        XCTAssertEqual(entry.value, "foo")
+    }
+
+    func testQueryingMetadata() throws {
+        // GIVEN
+            sut.log(level: .debug, message: "a", metadata: ["system": "auth"])
+            flush(store: store)
+
+        // WHEN
+        let request = NSFetchRequest<MessageEntity>(entityName: "MessageEntity")
+        request.predicate = NSPredicate(format: "text == %@ AND SUBQUERY(metadata, $entry, $entry.key == %@ AND $entry.value == %@).@count > 0", "a", "system", "auth")
+        let messages = try store.container.viewContext.fetch(request)
+
+        // THEN
+        let message = try XCTUnwrap(messages.first)
+        XCTAssertEqual(message.metadata.count, 1)
+        let entry = try XCTUnwrap(message.metadata.first)
+        XCTAssertEqual(message.text, "a")
+        XCTAssertEqual(entry.key, "system")
+        XCTAssertEqual(entry.value, "auth")
+    }
+}
+
+extension LogHandler {
+    func log(level: Logger.Level, message: Logger.Message, metadata: Logger.Metadata? = nil, file: String = #file, function: String = #function, line: UInt = #line) {
+        self.log(level: level, message: message, metadata: metadata, file: file, function: function, line: line)
     }
 }